Remove unmaintained compiler targets
Remove Parable and Retro Forth (retro11, retro12).

Rename the pl0_ansforth.py to pl0_compiler_forth.py to make it more
clear that this is using an alternative target/backend for the
compiler.

Remove the VARIANTS.md file. History should be present in the git log.

Add myself to the Copyright notice in README.md.

TODO: compile error for 41_recursion.pl0